--- apps/admin/src/service/common.ts
@@ -1,3 +1,0 @@
-export interface ApiResponse<T> {
-  data: T;
-}
